Fixed server tests
