+++ server/src/main/kotlin/com/filamagenta/endpoint/UserTransactionListOtherEndpoint.kt
@@ -1,6 +1,5 @@
 package com.filamagenta.endpoint
 
-import KoverIgnore
 import com.filamagenta.database.database
 import com.filamagenta.database.entity.Transaction
 import com.filamagenta.database.entity.User
@@ -8,42 +7,18 @@ import com.filamagenta.database.table.Transactions
 import com.filamagenta.endpoint.model.SecureEndpoint
 import com.filamagenta.endpoint.model.respondFailure
 import com.filamagenta.endpoint.model.respondSuccess
-import data.TransactionType
+import data.SerializableTransaction
 import io.ktor.server.application.ApplicationCall
 import io.ktor.server.application.call
 import io.ktor.server.util.getValue
 import io.ktor.util.pipeline.PipelineContext
-import kotlinx.serialization.Serializable
+import java.time.ZoneId
+import java.time.format.DateTimeFormatter
 import response.Errors
+import response.endpoint.UserTransactionListResult
 import server.Endpoints
 
 object UserTransactionListOtherEndpoint : SecureEndpoint(Endpoints.User.Transactions.ListOther) {
-    @KoverIgnore
-    @Serializable
-    data class UserTransactionsResponse(
-        val transactions: List<SerializableTransaction>
-    ) {
-        @KoverIgnore
-        @Serializable
-        data class SerializableTransaction(
-            val date: String,
-            val description: String,
-            val income: Boolean,
-            val units: UInt,
-            val pricePerUnit: Float,
-            val type: TransactionType
-        ) {
-            constructor(transaction: Transaction) : this(
-                transaction.date.toString(),
-                transaction.description,
-                transaction.income,
-                transaction.units,
-                transaction.pricePerUnit,
-                transaction.type
-            )
-        }
-    }
-
     override suspend fun PipelineContext<Unit, ApplicationCall>.secureBody(user: User) {
         val userId: Int by call.parameters
 
@@ -55,8 +30,21 @@ object UserTransactionListOtherEndpoint : SecureEndpoint(Endpoints.User.Transact
         }
 
         respondSuccess(
-            data = UserTransactionsResponse(
-                transactions = transactions.map { UserTransactionsResponse.SerializableTransaction(it) }
+            data = UserTransactionListResult(
+                transactions = transactions.map {
+                    with(it) {
+                        SerializableTransaction(
+                            id.value,
+                            date.atStartOfDay(ZoneId.systemDefault()).format(DateTimeFormatter.ISO_DATE_TIME),
+                            description,
+                            income,
+                            units,
+                            pricePerUnit,
+                            type,
+                            user.id.value
+                        )
+                    }
+                }
             )
         )
     }
